requirements.txt
flask==3.0.2
python-dotenv==1.0.1
agno==0.1.0
openai==1.12.0
gunicorn==21.2.0 